Error Handling - BE logic + FE notifications

+++ src/main/java/com/example/fullstackproject/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.fullstackproject.student;
 
+import com.example.fullstackproject.student.exception.BadRequestException;
+import com.example.fullstackproject.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,11 +16,22 @@ public class StudentService {
     public List<Student> getAllStudents() {
         return studentRepository.findAll();
     }
+
     public void addStudent(Student student) {
+        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException("Email " + student.getEmail() + " taken");
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists"
+            );
+        }
         studentRepository.deleteById(studentId);
     }
 }
